chore(refactor): fixed pathing for lib
