+++ agents/components/LogOut.js
@@ -3,7 +3,7 @@ import FlatButton from 'material-ui/FlatButton'
 import { connect as connectFela } from 'react-fela'
 import { pipe } from 'ramda'
 
-import { FormattedMessage } from '../../../cobuy/lib/Intl'
+import { FormattedMessage } from '../../lib/Intl'
 import styles from '../styles/LogOut'
 
 function LogOut (props) {
