+++ agents/components/Profile.js
@@ -5,7 +5,7 @@ import { Field, reduxForm as connectForm } from 'redux-form'
 import { pipe } from 'ramda'
 import { TextField } from 'redux-form-material-ui'
 
-import { FormattedMessage } from '../../../cobuy/lib/Intl'
+import { FormattedMessage } from '../../lib/Intl'
 import styles from '../styles/Profile'
 import Button from '../../app/components/Button'
 import AvatarField from '../../app/components/AvatarField'
